[fix][x86] compliation warnings

+++ src/ppl/nn/engines/x86/kernels/onnx/prelu_kernel.cc
@@ -21,7 +21,6 @@
 namespace ppl { namespace nn { namespace x86 {
 
 ppl::common::RetCode PReluKernel::DoExecute(KernelExecContext* ctx) {
-
     PPLNN_X86_REQUIRED_INPUT(X, 0);
     PPLNN_X86_REQUIRED_INPUT(slope, 1);
     PPLNN_X86_REQUIRED_OUTPUT(Y, 0);
@@ -38,41 +37,42 @@ ppl::common::RetCode PReluKernel::DoExecute(KernelExecContext* ctx) {
     const auto data_type = X->GetShape().GetDataType();
     const auto data_format = X->GetShape().GetDataFormat();
 
-    if (data_format == ppl::common::DATAFORMAT_NDARRAY){
+    if (data_format == ppl::common::DATAFORMAT_NDARRAY) {
         if (slope->GetShape().GetDimCount() > X->GetShape().GetDimCount()) {
             LOG(ERROR) << "prelu slope dimcount is bigger than input dimcount.";
             return ppl::common::RC_UNSUPPORTED;
         }
         auto channels = X->GetShape().GetDimCount() > 1 ? X->GetShape().GetDim(1) : 1;
-        if (slope->GetShape().GetElementsExcludingPadding() != channels) {
+        if (slope->GetShape().GetElementsExcludingPadding() != (uint64_t)channels) {
             LOG(ERROR) << "prelu only support channel broadcasting.";
             return ppl::common::RC_UNSUPPORTED;
         }
         if (slope->GetShape().GetDimCount() > 1 && channels > 1) {
             int32_t broadcast_dim = -1;
-            for(uint32_t i = 0; i < slope->GetShape().GetDimCount(); ++i) {
-                if (slope->GetShape().GetDim(i) == channels) { broadcast_dim = i; break; }
+            for (uint32_t i = 0; i < slope->GetShape().GetDimCount(); ++i) {
+                if (slope->GetShape().GetDim(i) == channels) {
+                    broadcast_dim = i;
+                    break;
+                }
             }
-            if (broadcast_dim == -1
-                || slope->GetShape().GetDimCount() - broadcast_dim != X->GetShape().GetDimCount() - 1) {
+            if (broadcast_dim == -1 ||
+                slope->GetShape().GetDimCount() - broadcast_dim != X->GetShape().GetDimCount() - 1) {
                 LOG(ERROR) << "prelu only support channel broadcasting.";
                 return ppl::common::RC_UNSUPPORTED;
             }
         }
     } else {
-        LOG(ERROR) << "unsupported data format: " << ppl::common::GetDataFormatStr(data_format) << "."; 
+        LOG(ERROR) << "unsupported data format: " << ppl::common::GetDataFormatStr(data_format) << ".";
         return ppl::common::RC_UNSUPPORTED;
     }
 
     if (data_type == ppl::common::DATATYPE_FLOAT32) {
         if (MayUseISA(ppl::common::ISA_X86_AVX)) {
             return kernel::x86::prelu_channel_shared_fp32_avx(&X->GetShape(), X->GetBufferPtr<float>(),
-                                                    slope->GetBufferPtr<float>(),
-                                                    Y->GetBufferPtr<float>());
+                                                              slope->GetBufferPtr<float>(), Y->GetBufferPtr<float>());
         } else if (MayUseISA(ppl::common::ISA_X86_SSE)) {
             return kernel::x86::prelu_channel_shared_fp32_sse(&X->GetShape(), X->GetBufferPtr<float>(),
-                                                    slope->GetBufferPtr<float>(),
-                                                    Y->GetBufferPtr<float>());
+                                                              slope->GetBufferPtr<float>(), Y->GetBufferPtr<float>());
         } else {
             LOG(ERROR) << "get unsupported isa " << GetISA();
         }
